+++ plugins/hls-refine-imports-plugin/test/testdata/WithOverride.expected.hs
@@ -3,9 +3,10 @@ module Main where
 import B ( b1 )
 import C ( c1 )
 import D
+import F
 import Data.List (intercalate)
 
 main :: IO ()
 main = putStrLn 
      $ "hello " 
-    <> intercalate ", " [b1, c1, e1]
+    <> intercalate ", " [b1, c1, e1, f1, g1]