+++ plugins/hls-refine-imports-plugin/test/testdata/WithOverride.hs
@@ -2,9 +2,10 @@ module Main where
 
 import A
 import D
+import F
 import Data.List (intercalate)
 
 main :: IO ()
 main = putStrLn 
      $ "hello " 
-    <> intercalate ", " [b1, c1, e1]
+    <> intercalate ", " [b1, c1, e1, f1, g1]